Fix google drive account add and manage ui update

+++ api/src/controllers/connector/connector.controller.js
@@ -3,6 +3,11 @@ const clog = require("../../services/ChalkService");
 const { commonConfig } = require("../../config/config");
 const { generateFbAuthUrl } = require("./facebook.controller");
 const fs = require("fs");
+const { getAccountDatabyId } = require("../../services/accountService");
+const {
+  getValidGoogleAccessToken,
+} = require("../../services/connectorService");
+
 // Function will return the callback data and query to frontend
 exports.connectCallbackController = async (req, res) => {
   let code = req?.query?.code;
@@ -16,7 +21,6 @@ exports.connectCallbackController = async (req, res) => {
   }
   return res.json({ body: req?.body, query: req?.query, credentials });
 };
-console.log("Oauth Cred", commonConfig.connector.google);
 
 const oauth2Client = new google.auth.OAuth2(
   commonConfig.connector.google.clientId,
@@ -24,6 +28,7 @@ const oauth2Client = new google.auth.OAuth2(
   commonConfig.connectCallbackUrl
 );
 
+// Function to get google creds
 exports.getGoogleAccessTokenfromAuthCode = async (auth_code) => {
   try {
     console.log("auth_code", auth_code);
@@ -84,18 +89,35 @@ const getDriveObject = (access_token) => {
   return drive;
 };
 
+const getDriveObjectFromAccountId = async (accountId) => {
+  try {
+    let {
+      status,
+      metadata: accountMetadata,
+      message,
+    } = await getAccountDatabyId(accountId);
+    if (!status) {
+      return { status: false, message: message ?? "Drive Account not found" };
+    }
+    let { access_token } = await getValidGoogleAccessToken(
+      accountMetadata?.credentials
+    );
+    const drive = getDriveObject(access_token);
+    return { status: true, drive, message: "Drive Account object created" };
+  } catch (Err) {
+    clog.error(Err);
+    return { status: false, message: Err?.message };
+  }
+};
+
 // API to connect Google Drive account
 exports.connectGDriveAccount = async (req, res) => {
   try {
-    const youtubeAuthurl = oauth2Client.generateAuthUrl({
+    const driveAuthUrl = oauth2Client.generateAuthUrl({
       access_type: "offline",
       scope: commonConfig.connector.drive.scopes,
     });
-    return res.REST.SUCCESS(
-      1,
-      "Auth url generated successfully",
-      youtubeAuthurl
-    );
+    return res.REST.SUCCESS(1, "Auth url generated successfully", driveAuthUrl);
   } catch (Err) {
     clog.error(Err);
     let errorMessage = Err?.message;
@@ -106,10 +128,13 @@ exports.connectGDriveAccount = async (req, res) => {
 // API to get items list from Google Drive
 exports.getGdriveItemsList = async (req, res) => {
   try {
-    const { access_token, folderId } = req.body;
-    oauth2Client.setCredentials({ access_token });
-    const drive = google.drive({ version: "v3", auth: oauth2Client });
-
+    const { storageAccountId, folderId } = req.body;
+    const { status, drive, message } = await getDriveObjectFromAccountId(
+      storageAccountId
+    );
+    if (!status) {
+      return res.REST.BADREQUEST(0, message);
+    }
     // Determine parent folder (root or specified folder)
     const parentFolder = folderId ? `'${folderId}'` : `'root'`;
 
@@ -145,12 +170,18 @@ exports.getGdriveItemsList = async (req, res) => {
 // API to get file data by ID
 exports.getGdriveFileData = async (req, res) => {
   try {
-    const { access_token, fileId } = req.body;
+    const { storageAccountId, fileId } = req.body;
+
     if (!fileId) {
       return res.REST.BADREQUEST(0, "File ID is required");
     }
-    oauth2Client.setCredentials({ access_token });
-    const drive = google.drive({ version: "v3", auth: oauth2Client });
+
+    const { status, drive, message } = await getDriveObjectFromAccountId(
+      storageAccountId
+    );
+    if (!status) {
+      return res.REST.BADREQUEST(0, message);
+    }
     // Fetch file metadata
     const response = await drive.files.get({
       fileId,
@@ -168,13 +199,16 @@ exports.getGdriveFileData = async (req, res) => {
 // API to delete a file by ID
 exports.deleteGdriveFile = async (req, res) => {
   try {
-    const { access_token, fileId } = req.query;
-    console.log(req.query);
+    const { storageAccountId, fileId } = req.query;
     if (!fileId) {
       return res.REST.BADREQUEST(0, "File ID is required");
     }
-    oauth2Client.setCredentials({ access_token });
-    const drive = google.drive({ version: "v3", auth: oauth2Client });
+    const { status, drive, message } = await getDriveObjectFromAccountId(
+      storageAccountId
+    );
+    if (!status) {
+      return res.REST.BADREQUEST(0, message);
+    }
     // Delete the file
     await drive.files.delete({ fileId });
 
@@ -185,29 +219,19 @@ exports.deleteGdriveFile = async (req, res) => {
   }
 };
 
-// **** Manage GDrive Connector ******
-// API to connect Google Drive account
-exports.connectFbAccount = async (req, res) => {
-  try {
-    const fbAuthUrl = generateFbAuthUrl();
-    return res.REST.SUCCESS(1, "Auth url generated successfully", fbAuthUrl);
-  } catch (Err) {
-    clog.error(Err);
-    let errorMessage = Err?.message;
-    return res.REST.SERVERERROR(0, errorMessage, Err);
-  }
-};
 // API to add folder
 exports.addGdriveFolder = async (req, res) => {
   try {
-    const { folderName, access_token, parentFolderId } = req.body;
-    console.log("reqbody", req?.body);
+    const { folderName, storageAccountId, parentFolderId } = req.body;
     if (!folderName) {
       return res.REST.BADREQUEST(0, "Folder name is required");
     }
-
-    // Authenticate with Google Drive API (assumes OAuth2 setup)
-    const drive = getDriveObject(access_token);
+    const { status, drive, message } = await getDriveObjectFromAccountId(
+      storageAccountId
+    );
+    if (!status) {
+      return res.REST.BADREQUEST(0, message);
+    }
 
     // Create the folder
     const folderMetadata = {
@@ -232,13 +256,17 @@ exports.addGdriveFolder = async (req, res) => {
 // API to add file
 exports.uploadGdriveFile = async (req, res) => {
   try {
-    const { folderId, access_token } = req.body;
+    const { folderId, storageAccountId } = req.body;
 
     if (!req.file) {
       return res.REST.BADREQUEST(0, "No file uploaded");
     }
-
-    const drive = getDriveObject(access_token);
+    const { status, drive, message } = await getDriveObjectFromAccountId(
+      storageAccountId
+    );
+    if (!status) {
+      return res.REST.BADREQUEST(0, message);
+    }
 
     const fileMetadata = {
       name: req.file.originalname,
@@ -271,7 +299,7 @@ exports.uploadGdriveFile = async (req, res) => {
 // To Rename a gdrive file or folder
 exports.renameGdriveItem = async (req, res) => {
   try {
-    const { fileId, newName, access_token } = req.body;
+    const { fileId, newName, storageAccountId } = req.body;
 
     if (!fileId || !newName) {
       return res.REST.BADREQUEST(
@@ -279,9 +307,15 @@ exports.renameGdriveItem = async (req, res) => {
         "Missing required parameters: fileId and newName"
       );
     }
-    const driveService = getDriveObject(access_token);
 
-    const updatedFile = await driveService.files.update({
+    const { status, drive, message } = await getDriveObjectFromAccountId(
+      storageAccountId
+    );
+    if (!status) {
+      return res.REST.BADREQUEST(0, message);
+    }
+
+    const updatedFile = await drive.files.update({
       fileId: fileId,
       resource: { name: newName },
       fields: "id, name",
@@ -294,6 +328,18 @@ exports.renameGdriveItem = async (req, res) => {
   } catch (err) {
     console.error("Google Drive API Error:", err);
     return res.REST.SERVERERROR(0, err?.message || "An error occurred", err);
+  }
+};
+
+// Manage Facebook Connector
+exports.connectFbAccount = async (req, res) => {
+  try {
+    const fbAuthUrl = generateFbAuthUrl();
+    return res.REST.SUCCESS(1, "Auth url generated successfully", fbAuthUrl);
+  } catch (Err) {
+    clog.error(Err);
+    let errorMessage = Err?.message;
+    return res.REST.SERVERERROR(0, errorMessage, Err);
   }
 };
